move some files and remove console logs
