+++ BlazarUI/app/stylus/components/branch-build-history.styl
@@ -1,17 +1,2 @@
 .branch-build-history-row
   cursor pointer
-
-.malformed-files__alert-wrapper
-  margin-top 8px
-
-.malformed-files__alert
-  cursor pointer
-  
-  &:hover
-    color darkred
-    
-.malformed-files__file
-  padding 10px
-  
-.malformed-files__file-contents
-  margin-top 5px
