+++ BlazarUI/app/stylus/components/malformed-file-notification.styl
@@ -1,0 +1,14 @@
+.malformed-files__alert-wrapper
+  margin-top 8px
+
+.malformed-files__alert
+  cursor pointer
+  
+  &:hover
+    color darkred
+    
+.malformed-files__file
+  padding 10px
+  
+.malformed-files__file-contents
+  margin-top 5px
