+++ BlazarUI/app/stylus/main.styl
@@ -32,6 +32,7 @@
 @import 'components/icon'
 @import 'components/json-modal'
 @import 'components/loader'
+@import 'components/malformed-file-notification'
 @import 'components/module-modal'
 @import 'components/muted-message'
 @import 'components/pagination'
